Removed crossbeam and used mpsc from tokio instead.
I learned today, looping while waiting for an even on a slow stream is
called "busy waiting." At first I felt attacked but yep, that makes
sense.

I've replaced this "hot" codepath which was eating my 22 core CPU up
with a simple solution from tokio- unbounded_channel and a while loop
with an await. Calmly wait for messages on the channel and do the
action.

Result? 0% CPU for 99% of the time instead of 78% CPU all the time.